+++ packages/scene/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @glissade/scene
 
+## 0.73.0-pre.0
+
+### Patch Changes
+
+- @glissade/core@0.73.0-pre.0
+
 ## 0.72.0
 
 ### Patch Changes